All test bench are finish.
I just completed ALU and its test bench, and also complete for the
child component of ALU: Computational, Flag Register, Arithmetic, Logic
and Shift. Hardware Access Unit test bench is finish.

There are some problem before with the TEST value for Logic Unit and
also other function in Arithmetic Unit. But I have already solved it.

All that is left is for us to do is the processor and the documentation.